+++ build-tools/documenter-new/src/component-definition.ts
@@ -2,28 +2,16 @@
 // SPDX-License-Identifier: Apache-2.0
 import ts from 'typescript';
 
-import { extractDeclaration, isOptional, stringifyType } from './type-utils.js';
+import type { ExpandedProp } from './extractor';
 import type {
   ComponentDefinition,
   ComponentFunction,
   ComponentProperty,
   ComponentRegion,
   EventHandler,
-  FunctionDefinition,
-  ObjectDefinition,
-  UnionTypeDefinition,
-} from './types.js';
-
-export interface ExpandedProp {
-  name: string;
-  type: string;
-  isOptional: boolean;
-  rawType: ts.Type;
-  description: {
-    text: string;
-    tags: Array<{ name: string; text: string | undefined }>;
-  };
-}
+} from './interfaces';
+import { getObjectDefinition } from './object-definition';
+import { extractDeclaration, stringifyType } from './type-utils';
 
 function getCommentTag(property: ExpandedProp, name: string) {
   const tag = property.description.tags.find(tag => tag.name === name);
@@ -81,7 +69,7 @@ export function buildComponentDefinition(
       (property): ComponentProperty => ({
         name: property.name,
         type: property.type,
-        inlineType: getObjectDefinition(property.name, property.rawType, checker),
+        inlineType: getObjectDefinition(property.rawType, checker),
         optional: property.isOptional,
         description: property.description.text,
         defaultValue: defaultValues[property.name],
@@ -92,7 +80,7 @@ export function buildComponentDefinition(
       })
     ),
     events: events.map((event): EventHandler => {
-      const { detailType, detailInlineType, cancelable } = extractEventDetails(event.name, event.rawType, checker);
+      const { detailType, detailInlineType, cancelable } = extractEventDetails(event.rawType, checker);
       return {
         name: event.name,
         description: event.description.text,
@@ -105,108 +93,18 @@ export function buildComponentDefinition(
   };
 }
 
-// TODO future feature
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-function stringifyFlags(type: ts.Type) {
-  const flags = [];
-  for (const [key, value] of Object.entries(ts.TypeFlags)) {
-    if (typeof value !== 'number') {
-      continue;
-    }
-    if (type.flags & value) {
-      flags.push(key);
-    }
-  }
-  return flags;
-}
-
-function isArrayType(type: ts.Type) {
-  const symbol = type.getSymbol();
-  if (!symbol) {
-    return false;
-  }
-  return symbol.getName() === 'Array' || symbol.getName() === 'ReadonlyArray';
-}
-
-function getObjectDefinition(
-  name: string,
-  type: ts.Type,
-  checker: ts.TypeChecker
-): ObjectDefinition | FunctionDefinition | UnionTypeDefinition | undefined {
+function extractEventDetails(type: ts.Type, checker: ts.TypeChecker) {
   const realType = type.getNonNullableType();
-  const realTypeName = stringifyType(realType, checker);
-  if (
-    realType.flags & ts.TypeFlags.String ||
-    realType.flags & ts.TypeFlags.StringLiteral ||
-    realType.flags & ts.TypeFlags.Boolean ||
-    realType.flags & ts.TypeFlags.Number ||
-    isArrayType(realType) ||
-    realTypeName === 'HTMLElement'
-  ) {
-    // do not expand built-in Javascript methods on primitive values
-    return;
+  const handlerName = realType.aliasSymbol?.getName();
+  if (handlerName !== 'CancelableEventHandler' && handlerName !== 'NonCancelableEventHandler') {
+    throw new Error(`Unknown event handler type: ${checker.typeToString(realType)}`);
   }
-  if (realType.isUnionOrIntersection()) {
-    if (!realType.types.every(subtype => subtype.isStringLiteral() || subtype.isNumberLiteral())) {
-      // only print out enum types of primitives
-      return;
-    }
-    return {
-      name: realTypeName,
-      type: 'union',
-      values: realType.types.map(subtype => (subtype as ts.StringLiteralType).value),
-    };
-  }
-  if (realType.getProperties().length > 0) {
-    return {
-      name: realTypeName,
-      type: 'object',
-      properties: realType
-        .getProperties()
-        .map(prop => {
-          const propType = checker.getTypeAtLocation(extractDeclaration(prop));
-          return {
-            name: prop.getName(),
-            type: stringifyType(propType, checker),
-            optional: isOptional(propType),
-          };
-        })
-        .sort((a, b) => a.name.localeCompare(b.name)),
-    };
-  }
-  if (realType.getCallSignatures().length > 0) {
-    if (realType.getCallSignatures().length > 1) {
-      throw new Error('Multiple call signatures are not supported');
-    }
-    const signature = realType.getCallSignatures()[0];
-
-    return {
-      name: realTypeName,
-      type: 'function',
-      returnType: stringifyType(signature.getReturnType(), checker),
-      parameters: signature.getParameters().map(param => {
-        const paramType = checker.getTypeAtLocation(extractDeclaration(param));
-        return {
-          name: param.getName(),
-          type: stringifyType(paramType, checker),
-        };
-      }),
-    };
-  }
-  return;
-}
-
-function extractEventDetails(name: string, type: ts.Type, checker: ts.TypeChecker) {
-  const realType = type.getNonNullableType();
-  const cancelable = realType.aliasSymbol?.getName() === 'CancelableEventHandler';
-  // if (name === 'onActiveHrefChange') {
-  //   debugger;
-  // }
+  const cancelable = handlerName === 'CancelableEventHandler';
   const detailType = realType.aliasTypeArguments?.[0];
   if (detailType && detailType.getProperties().length > 0) {
     return {
       detailType: stringifyType(detailType, checker),
-      detailInlineType: getObjectDefinition(name, detailType, checker),
+      detailInlineType: getObjectDefinition(detailType, checker),
       cancelable,
     };
   }
