+++ build-tools/documenter-new/src/interfaces.ts
